Module iota_move_natives_latest::test_scenario

Structs

InMemoryTestStore

An in-memory test store is a thin wrapper around the in-memory storage in a mutex. The mutex allows this to be used by both the object runtime (for reading) and the test scenario (for writing) while hiding mutability.

Functions

allocate_receiving_ticket_for_object

deallocate_receiving_ticket_for_object

end_transaction

ids_for_address

most_recent_id_for_address

most_recent_id_shared

most_recent_immutable_id

take_from_address_by_id

take_immutable_by_id

take_shared_by_id

was_taken_from_address

was_taken_immutable

was_taken_shared